removed redundant type casts

+++ src/main/java/de/fhconfig/android/binding/viewAttributes/view/ConditionalAnimationTrigger.java
@@ -95,7 +95,7 @@ public class ConditionalAnimationTrigger extends AnimationTrigger implements Obs
 
 	public void onPropertyChanged(IObservable<?> prop,
 	                              Collection<Object> initiators) {
-		setAnimationId((int) mId.get());
+		setAnimationId(mId.get());
 		TriggerValue = mTrigger.get();
 		Type = mType.get();
 		TriggerCondition = mCondition.get();
